pyproject.toml
[tool.poetry]
name = "Talk-to-your-data Slackbot"
version = "0.1.0"
description = "An AI Slackbot that can answer questions about your data."
authors = ["Saif Ur Rehman <saifurrehmanakhund@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^2.0.3"
numpy = "^1.24.3"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
pyyaml = "^6.0.0"
jupyter = "^1.0.0"
ipykernel = "^6.0.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
pandasai = "^3.0.0"
pandasai-litellm = "^0.0.1"
python-dotenv = "^1.0.0"
pandasai-sql = "^0.1.7"
slack-bolt = "^1.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
